# src/apps/onprem/publication-api/openapi.yaml
openapi: 3.1.0

# 문법 명세(경로·필드·타입) 층 — 2층 구조의 하위(TODO §5).
# 시맨틱 계약(파라미터 의미·에러 시맨틱·식별자 정책)의 SSOT 는
# ../../../../docs/contracts/publication-api.md 이며, 이 문서는 그 하위 문법 층이다.
# 응답 필드 표기는 snake_case (ExplanationResponse @JsonNaming SnakeCase).
# 대외 산출물 — 표현은 writing-rules 준수(투자권유·예측이 아닌 "변동 요인 후보" 톤).
info:
  title: Publication API
  version: 1.0.0
  description: >-
    MTS 위젯이 직접 호출하는 조회 표면(ADR-0053). 특정 ETF·거래일의 **검수 완료 노출 문구**
    (변동 요인 후보 설명)를 반환한다 — 투자권유·예측이 아니다. 고객 식별은 어떤 형태로도 받지
    않는다(고객 해시 폐지 — ADR-0053, ADR-0013 대체). 인증 없는 공개 읽기 표면이며 남용 통제
    (rate limit·인증 헤더 strip)는 증권사 엣지가 전제한다. 시맨틱 계약 SSOT 는
    ../../../../docs/contracts/publication-api.md.

servers:
  - url: /
    description: publication-api 서비스 루트 (온프렘)

# 벤더 API Key 없음 — 인증 없는 공개 읽기 표면이고 남용 통제는 증권사 엣지 소관이다
# (publication-api.md·ADR-0053). 그래서 security 요구를 두지 않는다.
security: []

paths:
  /api/v1/explanations/{etfTicker}:
    get:
      operationId: getExplanation
      summary: ETF·거래일의 검수 완료 설명 조회
      description: >-
        해당 ETF 종목코드의 **노출 가능(Published) 설명**을 반환한다. `trade_date` 생략 시
        최신 거래일 게시분(게시 시각이 아니라 거래일 기준 — 과거일 검수분이 늦게 게시돼도 최신
        거래일이 이긴다). 노출 가능한 설명이 없으면 204(정상 — 모든 ETF 가 매일 설명을 갖지
        않는다). 노출 이력은 기록하지 않는다(ADR-0053) — 서버 기록은 요청 메트릭뿐이다.
      parameters:
        - name: etfTicker
          in: path
          required: true
          description: 국내 상장 ETF 종목코드(MVP 커버리지 — ADR-0024). 미상장 코드는 404.
          schema:
            type: string
          example: "069500"
        - name: trade_date
          in: query
          required: false
          description: >-
            조회 거래일. 생략하거나 공백뿐인 값(`trade_date=`·공백 문자열)이면 생략으로 처리해
            최신 거래일 게시분을 반환한다(서버 `parseTradeDate` 의 `isBlank` 분기). 공백 아닌
            값을 지정하면 yyyy-MM-dd 형식이어야 하며, 아니면 400(SERV4004).
          schema:
            # 서버가 실제로 받는 집합을 그대로 전사 — 공백뿐 문자열(생략과 동치, isBlank) 또는
            # yyyy-MM-dd. 코드 생성기·게이트웨이가 이 집합을 거부하지 않게 스키마로 인코딩한다.
            type: string
            anyOf:
              - pattern: '^\s*$'
              - format: date
          example: "2026-07-15"
      responses:
        '200':
          description: 노출 가능한 설명 1건. 본문 표기 snake_case.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExplanationResponse'
        '204':
          description: 해당 ETF·거래일에 노출 가능한 설명이 없음(정상). 본문 없음.
        '400':
          description: >-
            요청 위반. 도메인 코드 — SERV4004(trade_date 형식 위반).
            SERV4001~4003(구 헤더 검증)은 ADR-0053 으로 폐지 — 번호는 재사용하지 않는다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidTradeDate:
                  summary: trade_date 형식 위반
                  value:
                    isSuccess: false
                    code: SERV4004
                    message: trade_date 는 yyyy-MM-dd 형식이어야 합니다.
        '404':
          description: 알 수 없는 ETF 종목코드(SERV4040).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                unknownEtf:
                  summary: 미상장 코드
                  value:
                    isSuccess: false
                    code: SERV4040
                    message: 알 수 없는 ETF 종목코드입니다.
        '5XX':
          description: 서버 일시 장애. 위젯은 폴백(직전 노출분 유지 등)을 권장한다. 내부 메시지는 노출하지 않는다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    ExplanationResponse:
      type: object
      description: >-
        고객에게 노출되는 검수 완료 설명. `summary`는 검수를 거친 **최종 노출 문구**(원본 AI 문구가
        아니라 검수 반영본)이며, `disclaimer`는 화면에 반드시 함께 노출한다(투자권유 아님 안내).
      required: [publication_id, etf, trade_date, summary, confidence_level, evidences, disclaimer, published_at, explanation_as_of]
      properties:
        publication_id:
          type: string
          description: 온프렘 발번 게시 식별자.
        etf:
          $ref: '#/components/schemas/EtfInfo'
        trade_date:
          type: string
          format: date
        summary:
          type: string
          description: 검수 완료 최종 노출 문구(변동 요인 후보 설명 — 투자권유·예측 아님).
        confidence_level:
          type: [string, 'null']
          description: 신뢰도 — 미산정 시 null.
          enum: [HIGH, MEDIUM, LOW, null]
        evidences:
          type: array
          description: 근거 뉴스/공시 문서 목록("관련 뉴스/공시"). 없으면 빈 배열.
          items:
            $ref: '#/components/schemas/EvidenceItem'
        disclaimer:
          type: string
          description: >-
            활성 정책 버전(`policy_version.disclaimer_text` — 콘솔 발행분)의 안내 문구(투자권유 아님)
            — 화면 필수 노출. 조회 시점 최신값이라 재발행 시 기존 게시분 조회에도 즉시 반영된다.
            활성 버전이 없는 구간(첫 발행 전)에는 콘솔과 같은 기본 문구가 실린다.
        published_at:
          type: string
          format: date-time
          description: 게시 시각(오프셋 포함 ISO-8601).
        explanation_as_of:
          type: string
          format: date-time
          description: >-
            스냅샷 기준시각(오프셋 포함 ISO-8601). 같은 거래일에 유효 스냅샷이 여럿이면
            이 값이 최신인 것이 서빙된다("유효 최신 승리" — ADR-0045). 최신 무효화 시
            재노출되는 직전 스냅샷의 기준 시점을 화면이 이 값으로 안다.
        content_as_of:
          type: [string, "null"]
          format: date-time
          description: >-
            콘텐츠 기준시각(오프셋 포함 ISO-8601) — 설명 본문이 서술하는 구간의 끝(ALPHA-918).
            explanation_as_of(생성 시각)와 달리 본문 내용의 시점을 말한다. 구형 게시분은
            null — 소비자는 explanation_as_of 로 폴백한다.

    EtfInfo:
      type: object
      required: [ticker, name]
      properties:
        ticker:
          type: string
        name:
          type: [string, 'null']
          description: ETF 표시명 — 종목 마스터 매칭이 없으면 null(키는 항상 존재).

    EvidenceItem:
      type: object
      description: >-
        근거 문서 하나 — 온프렘이 번들에서 받아 저장한 형상 그대로 서빙(ALPHA-395·497). 대외
        노출용 계약이라 내부 wire 스키마(event-bundle)와 별개로 여기 정의한다.
      required: [kind, title, source, published_at]
      properties:
        kind:
          type: string
          enum: [NEWS, DISCLOSURE]
          description: 근거 종류.
        title:
          type: [string, 'null']
          description: 헤드라인 — 없으면 null.
        source:
          type: string
          description: 출처(예 DART·언론사).
        published_at:
          type: [string, 'null']
          format: date-time
          description: 문서 게시 시각 — 없으면 null.

    ErrorResponse:
      type: object
      description: >-
        공통 응답 포맷(jvm-common ApiResponse) — 4xx/5xx 에만 적용된다. 성공(200) 본문은
        이 포맷으로 감싸지 않는다.
      required: [isSuccess, code, message]
      properties:
        isSuccess:
          type: boolean
          const: false
        code:
          type: string
          description: 도메인 코드(예 SERV4004) 또는 공통 코드(예 COMMON404).
          example: SERV4004
        message:
          type: string
          example: trade_date 는 yyyy-MM-dd 형식이어야 합니다.